slideshow/requirements.txt
# Podcast Slideshow Generator Requirements
# Install with: pip install -r requirements.txt

# Image processing
Pillow>=10.0.0

# Video processing and editing
moviepy>=1.0.3

# Audio metadata reading
mutagen>=1.47.0

# Additional dependencies for moviepy
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# Optional: Progress bars for better user experience
tqdm>=4.65.0